# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs
  description: >-
    A simpleJobs  API to test job creation, update, filtering, deletion and many more.
    
    
    Let's dive in!!!
  contact:
    name: Gegabox Technologies Limited
    url: gegabox.ug
    email: gegaboxtechnologies@gmail.com
  version: '1.0'
servers:
  - url: http://localhost:3000/api/v1
    variables: {}
paths:
...
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
        - jobs
      summary: Get single job
      description: Here provide a valid job id in order to get the single specified job
      operationId: Getsinglejob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    patch:
      tags:
        - jobs
      summary: Update a job
      description: Here provide a valid job id in order to update the job
      operationId: Updateajob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UpdateajobRequest'
                - example:
                    position: Flutter Programmer
                    status: interview
            example:
              position: Flutter Programmer
              status: interview
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    delete:
      tags:
        - jobs
      summary: Delete a job
      description: Here provide a valid job id in order to delete the job
      operationId: Deleteajob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
components:
  schemas:
...
    UpdateajobRequest:
      title: UpdateajobRequest
      required:
        - position
        - status
      type: object
      properties:
        position:
          type: string
        status:
          type: string
      example:
        position: Flutter Programmer
        status: interview
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...